# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RLEGS AM Monitoring & Bot Telegram API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/otp/request:
    post:
      operationId: otpRequest
      tags: [auth]
      summary: Request OTP for login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequestBody"
      responses:
        "200":
          description: OTP request successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpChallengeResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/otp/verify:
    post:
      operationId: otpVerify
      tags: [auth]
      summary: Verify OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyBody"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/otp/resend:
    post:
      operationId: otpResend
      tags: [auth]
      summary: Resend OTP
      responses:
        "200":
          description: OTP resent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpChallengeResponse"
        "400":
          description: No active challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        error:
          type: ["string", "null"]
        id:
          type: integer
        email:
          type: string
        role:
          type: string
        nextStep:
          type: string
          enum: [AUTHENTICATED, OTP_REQUIRED, TELEGRAM_LINK_REQUIRED]
        userId:
          type: integer
        nama:
          type: string
        tipe:
          type: ["string", "null"]
        telegramConnected:
          type: boolean
        presentationToken:
          type: string
          description: "Server-validated session token for presentation mode. Present only on successful presentation OTP verification."
      required: [id, email, role]
    OtpRequestBody:
      type: object
      properties:
        userId:
          type: integer
      required: [userId]
    OtpChallengeResponse:
      type: object
      properties:
        challengeId:
          type: string
        expiresAt:
          type: string
          format: date-time
      required: [challengeId, expiresAt]
    OtpVerifyBody:
      type: object
      properties:
        challengeId:
          type: string
        otp:
          type: string
        userId:
          type: ["integer", "null"]
      required: [challengeId, otp]
...